refact: rename `api` module names
They all start with archive. Will add more to this directory
